requirements.txt
Flask==3.0.2
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
Flask-APScheduler==1.13.1
pytz==2024.1 